Refactor integration test TestHostBuilder setup to shared abstract class (#512)
